+++ qplayer/include/state.h
@@ -44,7 +44,7 @@ class QState {
 private:
 	qsize_t index;
 	complex_t amplitude;
-	short stage;
+	size_t stage;
 
 public:
 	QState(qsize_t i, complex_t a);
@@ -57,8 +57,8 @@ public:
 	void setIndex(qsize_t i) { index = i; }
 	void setAmplitude(complex_t a) { amplitude = a; }
 	void resizeAmplitude(complex_t a) { amplitude /= a; }
-	void setStage(short s) { stage = s; }
-	short getStage(void) { return stage; }
+	void setStage(size_t s) { stage = s; }
+	size_t getStage(void) { return stage; }
 };
 
 #endif
